rename rag.main to rag.orchestrator
